+++ tests/unit/test_node.py
@@ -27,21 +27,22 @@ class TestMemcachedHostAddress:
 
 class TestNode:
     async def test_properties(self, connection_pool, memcached_host_address):
-        node = Node(memcached_host_address, 1, 60, 5, lambda _: _)
+        node = Node(memcached_host_address, 1, 1, 60, 5, lambda _: _)
         assert node.host == "localhost"
         assert node.port == 11211
         assert node.memcached_host_address == memcached_host_address
 
     async def test_str(self, connection_pool, memcached_host_address):
-        node = Node(memcached_host_address, 1, 60, 5, lambda _: _)
+        node = Node(memcached_host_address, 1, 1, 60, 5, lambda _: _)
         assert str(node) == "<Node host=localhost port=11211 closed=False>"
         assert repr(node) == "<Node host=localhost port=11211 closed=False>"
 
     async def test_connection_pool(self, connection_pool, memcached_host_address):
-        node = Node(memcached_host_address, 1, 60, 5, lambda _: _)
+        node = Node(memcached_host_address, 1, 1, 60, 5, lambda _: _)
         connection_pool.assert_called_with(
             memcached_host_address.address,
             memcached_host_address.port,
+            1,
             1,
             60,
             5,
@@ -52,12 +53,12 @@ class TestNode:
         connection_pool = Mock()
         connection_pool.close = CoroutineMock()
         mocker.patch("emcache.node.ConnectionPool", return_value=connection_pool)
-        node = Node(memcached_host_address, 1, 60, 5, lambda _: _)
+        node = Node(memcached_host_address, 1, 1, 60, 5, lambda _: _)
         await node.close()
         connection_pool.close.assert_called()
 
     async def test_connection_pool_healthiness_propagation(self, connection_pool, memcached_host_address):
         cb_mock = Mock()
-        node = Node(memcached_host_address, 1, 60, 5, cb_mock)
+        node = Node(memcached_host_address, 1, 1, 60, 5, cb_mock)
         node._on_connection_pool_healthy_status_change_cb(True)
         cb_mock.assert_called_with(node, True)